Back down C++20 experiment. Ahh well.
This reverts commit 7e19415754d360c74842a8d967d652e9d981bd7d.

+++ src/surge-xt/gui/widgets/XMLConfiguredMenus.cpp
@@ -322,7 +322,7 @@ void XMLMenuPopulator::populate()
 
     rootTree->buildJuceMenu(menu, this);
 
-    maxIdx = (int)allPresets.size();
+    maxIdx = allPresets.size();
 }
 
 OscillatorMenu::OscillatorMenu() : juce::Component(), WidgetBaseMixin<OscillatorMenu>(this)
